new test suite for QueryService class and all sublasses; comparison of results of HTML and Solr variants enabled and added
